Fixed background image not being tall enough when zoomed out or on large screen
Turns out, you need "height: 100%" applied to html, or body, or something. That way, when the div is set to 100%, it stretches down the entire page. Also, changed background-size to 100% 100%. Cover h1 height needed to be changed because its size was gigantic when I added the background image changes.

I may or may not have removed an unnecessary div. Not sure, but it shows up as a difference.